Extract short read assembler testing into separate file
