+++ test/image/test_execute.py
@@ -4,7 +4,6 @@ import biobox.image.execute as exe
 import biobox.image.volume  as vol
 import biobox.util          as util
 
-IMAGE = 'bioboxes/velvet@sha256:6611675a6d3755515592aa71932bd4ea4c26bccad34fae7a3ec1198ddcccddad'
 
 
 def test_prepare_biobox_file():
@@ -32,29 +31,3 @@ def test_prepare_volumes():
     for d in map(vol.get_host_path, volumes):
         assert os.path.isdir(d) == True
         assert os.access(d, os.W_OK)
-
-
-def test_create_container():
-    cnt = exe.create_container(
-            IMAGE,
-            hlp.biobox_args(),
-            tempfile.mkdtemp())
-    assert "Id" in cnt
-    hlp.clean_up_container(cnt["Id"])
-
-
-@pytest.mark.slow
-def test_executing_container():
-    out_dir = tempfile.mkdtemp()
-    cnt = exe.create_container(
-            IMAGE,
-            hlp.biobox_args(),
-            out_dir,
-            "default",
-            {"detach" : False})
-    id_ = cnt['Id']
-    util.client().start(id_)
-    util.client().wait(id_)
-    assert funcy.get_in(util.client().inspect_container(id_), ['State', 'ExitCode']) == 0
-    assert os.path.isfile(os.path.join(out_dir, 'contigs.fa'))
-    hlp.clean_up_container(id_)
